+++ src/libraries/GeKo_Graphics/Light/PointLight.cpp
@@ -9,7 +9,7 @@ PointLight::PointLight()
 	m_radius = 0;
 }
 
-PointLight::PointLight(glm::vec4 position, glm::vec4 color, bool isActive, double radius)
+PointLight::PointLight(glm::vec4 position, glm::vec4 color, bool isActive, float radius)
 {
 	m_position = position;
 	m_color = color;
